+++ src/domain/payment/ports/payment_adapter_interface.py
@@ -1,4 +1,5 @@
 import abc
+from decimal import Decimal
 
 from domain.payment.model.value_objects import PaymentId
 
@@ -7,7 +8,7 @@ class PaymentAdapterInterface(abc.ABC):
     """Abstraction for payment provider integration."""
 
     @abc.abstractmethod
-    async def new_payment(self, total_price: float) -> PaymentId:
+    async def new_payment(self, total_price: Decimal) -> PaymentId:
         """Create a new payment for the given total price."""
         raise NotImplementedError()
